LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/constraints/form-validation-validity-rangeOverflow-expected.txt


FAIL [INPUT in DATETIME-LOCAL status] The datetime-local type must be supported. assert_equals: The datetime-local type should be supported. expected "datetime-local" but got "text"
FAIL [INPUT in DATE status] The date type must be supported. assert_equals: The date type should be supported. expected "date" but got "text"
FAIL [INPUT in MONTH status] The month type must be supported. assert_equals: The month type should be supported. expected "month" but got "text"
FAIL [INPUT in WEEK status] The week type must be supported. assert_equals: The week type should be supported. expected "week" but got "text"
FAIL [INPUT in TIME status] The time type must be supported. assert_equals: The time type should be supported. expected "time" but got "text"
PASS [INPUT in NUMBER status] The max attribute is not set
PASS [INPUT in NUMBER status] Value is empty string
PASS [INPUT in NUMBER status] The max is greater than value(integer)
PASS [INPUT in NUMBER status] The max is greater than value(floating number)
PASS [INPUT in NUMBER status] The max equals to value
PASS [INPUT in NUMBER status] The value is not a number
FAIL [INPUT in NUMBER status] The value is greater than max(integer) assert_true: The validity.rangeOverflow should be true, when control is disabled. expected true got false
FAIL [INPUT in NUMBER status] The value is greater than max(floating number) assert_true: The validity.rangeOverflow should be true, when control is disabled. expected true got false
FAIL [INPUT in NUMBER status] The value is greater than max(special floating number) assert_true: The validity.rangeOverflow should be true, when control is disabled. expected true got false
FAIL [INPUT in NUMBER status] The value is greater than max(scientific notation) assert_true: The validity.rangeOverflow should be true, when control is disabled. expected true got false
